Split table input and output into two modules
Just copy and paste for now, so review of this commit alone is quite nice.
